text_layout: fix phantom line on wrap+hard-break; double vertical accumulation (elements #370)
Phantom-line bug (skia): a wrapped paragraph whose last line slightly overflows
was kept whole at end-of-text, but BROKEN when followed by a hard '\n' -- the
overflow check fired on the final glyph before the break, producing an extra
blank line. Suppress the overflow break when the next glyph forces a hard break,
so a '\n'-terminated segment's last line is kept whole like end-of-text. (Cairo
and Quartz/CoreText are already self-consistent here; verified by the per-edit
fuzz.)

Precision: accumulate vertical position in double across all three backends and
in text_layout_ex's paragraph offsets. Line height sums over thousands of rows;
float drifted ~0.2px by the bottom of a novel-length document.

Tests: the novel editing-session acid test now compares caret geometry and the
rendered editing window against a fresh text_layout_ex (same scheme, robust on
every backend), keeping the integer num_lines check against a single text_layout
for layout semantics. Passes on skia, cairo and quartz.

+++ lib/impl/cairo/text_layout.cpp
@@ -151,7 +151,8 @@ namespace cycfi::artist
 
       auto linfo = glf(0);
       float target = linfo.width;
-      float y = 0;
+      double y = 0;   // accumulate vertical position in double: line height sums
+                      // over thousands of rows, and float would drift ~0.2px.
       float x = 0;
       std::size_t glyph_start = 0;
       constexpr std::size_t NONE = std::size_t(-1);
@@ -197,7 +198,7 @@ namespace cycfi::artist
             }
 
             _rows.push_back(row_info{
-               glyph_start, n, linfo.offset, y, line_width,
+               glyph_start, n, linfo.offset, float(y), line_width,
                std::move(row_pos)});
 
             y += finfo.line_height;
@@ -265,7 +266,7 @@ namespace cycfi::artist
             ? positions[n-1] + _glyphs[glyph_start + n - 1].x_advance
             : 0;
          _rows.push_back(row_info{
-            glyph_start, n, linfo.offset, y, line_width,
+            glyph_start, n, linfo.offset, float(y), line_width,
             std::move(positions)});
       }
